build: room 테이블 수정

+++ docker/db/mysql/init/create_table.sql
@@ -1,13 +1,13 @@
 CREATE TABLE room
 (
-    room_id INT PRIMARY KEY AUTO_INCREMENT,
+    room_id BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
     name    VARCHAR(30) NOT NULL
 );
 
 CREATE TABLE piece
 (
-    piece_id      INT PRIMARY KEY AUTO_INCREMENT,
-    room          INT         NOT NULL,
+    piece_id      BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
+    room          BIGINT(20)  NOT NULL,
     name          VARCHAR(10) NOT NULL,
     team          VARCHAR(10) NOT NULL,
     position_file VARCHAR(10) NOT NULL,
